+++ InspirationBulb.h
@@ -7,7 +7,7 @@
 class InspirationBulb:public Tower
 {
 public:
-    explicit InspirationBulb(int range,QGraphicsItem* parent = nullptr);
+    explicit InspirationBulb(double range,QGraphicsItem* parent = nullptr);
     void upgrade()override;
 };
